EEvening Commit

+++ Blazr.Demo.Database.UI/Entities/WeatherForecast/Components/WeatherForecastListForm.razor.cs
@@ -1,9 +1,14 @@
-﻿using Blazr.Core.Toaster;
-
+﻿/// ============================================================
+/// Author: Shaun Curtis, Cold Elm Coders
+/// License: Use And Donate
+/// If you use it, donate something to a charity somewhere
+/// ============================================================
 namespace Blazr.Demo.Database.UI;
 
 public partial class WeatherForecastListForm : ComponentBase
 {
+    [CascadingParameter] public Task<AuthenticationState>? AuthTask { get; set; }
+
     [Inject] private WeatherForecastsViewService? listViewService { get; set; }
 
     [Inject] private WeatherForecastViewService? recordViewService { get; set; }
@@ -13,6 +18,10 @@ public partial class WeatherForecastListForm : ComponentBase
     [Inject] private ToasterService? toasterService { get; set; }
 
     [Inject] private ResponseMessageStore? ResponseMessageStore { get; set; }
+
+    [Inject] private IClientAuthenticationService? clientAuthenticationService { get; set; }
+
+    [Inject] protected IAuthorizationService? AuthorizationService { get; set; }
 
     private bool isLoading => listViewService!.Records is null;
 
@@ -30,6 +39,20 @@ public partial class WeatherForecastListForm : ComponentBase
 
     private async Task DeleteRecord(Guid Id)
     {
+        await this.recordViewService!.GetForecastAsync(Id);
+
+        if (this.recordViewService.Record is null)
+        {
+            toasterService?.AddToast(Toast.NewTTD("Delete", $"Delete Failed.  the record does not exist", MessageColour.Danger, 30));
+            return;
+        }
+
+        if (!await this.CheckAuthorization(this.recordViewService.Record, AppPolicies.IsEditor))
+        {
+            toasterService?.AddToast(Toast.NewTTD("Delete", $"Delete Failed.  You do not have permissions to delete this record", MessageColour.Danger, 30));
+            return;
+        }
+
         var transactionId = Guid.NewGuid();
         await this.recordViewService!.DeleteRecordAsync(transactionId, Id);
 
@@ -69,11 +92,13 @@ public partial class WeatherForecastListForm : ComponentBase
 
     private async Task AddRecordAsync()
     {
+        var user = clientAuthenticationService?.GetCurrentIdentity();
         var transactionId = Guid.NewGuid();
         var record = new DcoWeatherForecast
         {
             Date = DateTime.Now,
             Id = Guid.NewGuid(),
+            OwnerId = SessionTokenManagement.GetIdentityId(clientAuthenticationService?.GetCurrentIdentity()),
             Summary = "Balmy",
             TemperatureC = 14
         };
@@ -91,6 +116,13 @@ public partial class WeatherForecastListForm : ComponentBase
 
     private void OnListChanged(object? sender, EventArgs e)
         => this.InvokeAsync(this.StateHasChanged);
+
+    protected virtual async Task<bool> CheckAuthorization(DcoWeatherForecast record, string policy)
+    {
+        var state = await AuthTask!;
+        var result = await this.AuthorizationService!.AuthorizeAsync(state.User, record, policy);
+        return result.Succeeded;
+    }
 
     public void Dispose()
         => this.listViewService!.ListChanged -= this.OnListChanged;
